+++ idl/src/lib.rs
@@ -1,0 +1,6 @@
+//! Anchor IDL.
+
+pub mod types;
+
+#[cfg(feature = "build")]
+pub mod build;