+++ lang/syn/src/codegen/accounts/mod.rs
@@ -37,10 +37,7 @@ pub fn generate(accs: &AccountsStruct) -> proc_macro2::TokenStream {
 
     #[cfg(feature = "idl-build")]
     {
-        #![allow(warnings)]
-        let no_docs = crate::idl::build::get_no_docs();
-        let idl_build_impl =
-            crate::idl::build::gen_idl_build_impl_for_accounts_struct(&accs, no_docs);
+        let idl_build_impl = crate::idl::build::gen_idl_build_impl_accounts_struct(accs);
         return quote! {
             #ret
             #idl_build_impl