+++ lang/syn/src/idl/mod.rs
@@ -2,6 +2,3 @@ pub mod types;
 
 #[cfg(feature = "idl-build")]
 pub mod build;
-
-#[cfg(feature = "idl-parse")]
-pub mod parse;