+++ spl/src/lib.rs
@@ -31,3 +31,6 @@ pub mod metadata;
 
 #[cfg(feature = "memo")]
 pub mod memo;
+
+#[cfg(feature = "idl-build")]
+mod idl_build;