+++ ts/packages/anchor/src/coder/index.ts
@@ -1,4 +1,4 @@
-import { IdlEvent, IdlTypeDef } from "../idl.js";
+import { IdlEvent } from "../idl.js";
 import { Event } from "../program/event.js";
 
 export * from "./borsh/index.js";
@@ -36,14 +36,14 @@ export interface StateCoder {
 
 export interface AccountsCoder<A extends string = string> {
   encode<T = any>(accountName: A, account: T): Promise<Buffer>;
-  decode<T = any>(accountName: A, ix: Buffer): T;
-  decodeUnchecked<T = any>(accountName: A, ix: Buffer): T;
+  decode<T = any>(accountName: A, acc: Buffer): T;
+  decodeUnchecked<T = any>(accountName: A, acc: Buffer): T;
   memcmp(accountName: A, appendData?: Buffer): any;
-  size(idlAccount: IdlTypeDef): number;
+  size(accountName: A): number;
 }
 
 export interface InstructionCoder {
-  encode(ixName: string, ix: any, discriminator?: Buffer): Buffer;
+  encode(ixName: string, ix: any): Buffer;
 }
 
 export interface EventCoder {