+++ ts/packages/anchor/src/program/namespace/methods.ts
@@ -7,7 +7,13 @@ import {
   TransactionInstruction,
   TransactionSignature,
 } from "@solana/web3.js";
-import { Idl, IdlAccountItem, IdlAccounts, IdlTypeDef } from "../../idl.js";
+import {
+  Idl,
+  IdlInstructionAccount,
+  IdlInstructionAccountItem,
+  IdlInstructionAccounts,
+  IdlTypeDef,
+} from "../../idl.js";
 import Provider from "../../provider.js";
 import {
   AccountsGeneric,
@@ -66,19 +72,57 @@ export class MethodsBuilderFactory {
   }
 }
 
-export type PartialAccounts<A extends IdlAccountItem = IdlAccountItem> =
-  Partial<{
-    [N in A["name"]]: PartialAccount<A & { name: N }>;
-  }>;
+type ResolvedAccounts<
+  A extends IdlInstructionAccountItem = IdlInstructionAccountItem
+> = PartialUndefined<ResolvedAccountsRecursive<A>>;
 
-type PartialAccount<A extends IdlAccountItem> = A extends IdlAccounts
+type ResolvedAccountsRecursive<
+  A extends IdlInstructionAccountItem = IdlInstructionAccountItem
+> = OmitNever<{
+  [N in A["name"]]: ResolvedAccount<A & { name: N }>;
+}>;
+
+type ResolvedAccount<
+  A extends IdlInstructionAccountItem = IdlInstructionAccountItem
+> = A extends IdlInstructionAccounts
+  ? ResolvedAccountsRecursive<A["accounts"][number]>
+  : A extends NonNullable<Pick<IdlInstructionAccount, "address">>
+  ? never
+  : A extends NonNullable<Pick<IdlInstructionAccount, "pda">>
+  ? never
+  : A extends NonNullable<Pick<IdlInstructionAccount, "relations">>
+  ? never
+  : A extends { signer: true }
+  ? Address | undefined
+  : PartialAccount<A>;
+
+type PartialUndefined<
+  T,
+  P extends keyof T = {
+    [K in keyof T]: undefined extends T[K] ? K : never;
+  }[keyof T]
+> = Partial<Pick<T, P>> & Pick<T, Exclude<keyof T, P>>;
+
+type OmitNever<T extends Record<string, any>> = {
+  [K in keyof T as T[K] extends never ? never : K]: T[K];
+};
+
+export type PartialAccounts<
+  A extends IdlInstructionAccountItem = IdlInstructionAccountItem
+> = Partial<{
+  [N in A["name"]]: PartialAccount<A & { name: N }>;
+}>;
+
+type PartialAccount<
+  A extends IdlInstructionAccountItem = IdlInstructionAccountItem
+> = A extends IdlInstructionAccounts
   ? PartialAccounts<A["accounts"][number]>
-  : A extends { isOptional: true }
+  : A extends { optional: true }
   ? Address | null
   : Address;
 
 export function isPartialAccounts(
-  partialAccount: PartialAccount<IdlAccountItem>
+  partialAccount: any
 ): partialAccount is PartialAccounts {
   return (
     typeof partialAccount === "object" &&
@@ -87,7 +131,7 @@ export function isPartialAccounts(
   );
 }
 
-export function flattenPartialAccounts<A extends IdlAccountItem>(
+export function flattenPartialAccounts<A extends IdlInstructionAccountItem>(
   partialAccounts: PartialAccounts<A>,
   throwOnNull: boolean
 ): AccountsGeneric {
@@ -108,122 +152,132 @@ export function flattenPartialAccounts<A extends IdlAccountItem>(
   return toReturn;
 }
 
-type SplInterface =
-  | "spl_transfer_hook_interface::initialize_extra_account_metas"
-  | "spl_transfer_hook_interface::execute";
-
-export class MethodsBuilder<IDL extends Idl, I extends AllInstructions<IDL>> {
-  private readonly _accounts: AccountsGeneric = {};
+export class MethodsBuilder<
+  IDL extends Idl,
+  I extends AllInstructions<IDL>,
+  A extends I["accounts"][number] = I["accounts"][number]
+> {
+  private _accounts: AccountsGeneric = {};
   private _remainingAccounts: Array<AccountMeta> = [];
   private _signers: Array<Signer> = [];
   private _preInstructions: Array<TransactionInstruction> = [];
   private _postInstructions: Array<TransactionInstruction> = [];
   private _accountsResolver: AccountsResolver<IDL>;
-  private _autoResolveAccounts: boolean = true;
-  private _args: Array<any>;
-  private _discriminator?: Buffer;
+  private _resolveAccounts: boolean = true;
 
   constructor(
-    _args: Array<any>,
+    private _args: Array<any>,
     private _ixFn: InstructionFn<IDL>,
     private _txFn: TransactionFn<IDL>,
     private _rpcFn: RpcFn<IDL>,
     private _simulateFn: SimulateFn<IDL>,
     private _viewFn: ViewFn<IDL> | undefined,
-    _provider: Provider,
-    private _programId: PublicKey,
-    _idlIx: AllInstructions<IDL>,
-    _accountNamespace: AccountNamespace<IDL>,
-    _idlTypes: IdlTypeDef[],
-    _customResolver?: CustomAccountResolver<IDL>
+    provider: Provider,
+    programId: PublicKey,
+    idlIx: AllInstructions<IDL>,
+    accountNamespace: AccountNamespace<IDL>,
+    idlTypes: IdlTypeDef[],
+    customResolver?: CustomAccountResolver<IDL>
   ) {
-    this._args = _args;
     this._accountsResolver = new AccountsResolver(
       _args,
       this._accounts,
-      _provider,
-      _programId,
-      _idlIx,
-      _accountNamespace,
-      _idlTypes,
-      _customResolver
+      provider,
+      programId,
+      idlIx,
+      accountNamespace,
+      idlTypes,
+      customResolver
     );
   }
 
-  public args(_args: Array<any>): void {
-    this._args = _args;
-    this._accountsResolver.args(_args);
+  public args(args: Array<any>): void {
+    this._args = args;
+    this._accountsResolver.args(args);
   }
 
-  public async pubkeys(): Promise<
-    Partial<InstructionAccountAddresses<IDL, I>>
-  > {
-    if (this._autoResolveAccounts) {
-      await this._accountsResolver.resolve();
-    }
-    return this._accounts as unknown as Partial<
-      InstructionAccountAddresses<IDL, I>
-    >;
+  /**
+   * Set instruction accounts with account resolution.
+   *
+   * This method only accepts accounts that cannot be resolved.
+   *
+   * See {@link accountsPartial} for overriding the account resolution or
+   * {@link accountsStrict} for strictly specifying all accounts.
+   */
+  public accounts(accounts: ResolvedAccounts<A>) {
+    // @ts-ignore
+    return this.accountsPartial(accounts);
   }
 
-  public interface(splInterface: SplInterface): MethodsBuilder<IDL, I> {
-    if (
-      splInterface ===
-      "spl_transfer_hook_interface::initialize_extra_account_metas"
-    ) {
-      this._discriminator = Buffer.from([43, 34, 13, 49, 167, 88, 235, 235]);
-    } else if (splInterface === "spl_transfer_hook_interface::execute") {
-      this._discriminator = Buffer.from([105, 37, 101, 197, 75, 251, 102, 26]);
-    } else {
-      throw new Error(`Unsupported interface: ${splInterface}`);
-    }
-    return this;
-  }
-
-  public accounts(
-    accounts: PartialAccounts<I["accounts"][number]>
-  ): MethodsBuilder<IDL, I> {
-    this._autoResolveAccounts = true;
+  /**
+   * Set instruction accounts with account resolution.
+   *
+   * There is no functional difference between this method and {@link accounts}
+   * method, the only difference is this method allows specifying all accounts
+   * even if they can be resolved. On the other hand, {@link accounts} method
+   * doesn't accept accounts that can be resolved.
+   */
+  public accountsPartial(accounts: PartialAccounts<A>) {
+    this._resolveAccounts = true;
     this._accountsResolver.resolveOptionals(accounts);
     return this;
   }
 
-  public accountsStrict(
-    accounts: Accounts<I["accounts"][number]>
-  ): MethodsBuilder<IDL, I> {
-    this._autoResolveAccounts = false;
+  /**
+   * Set instruction accounts without account resolution.
+   *
+   * All accounts strictly need to be specified when this method is used.
+   *
+   * See {@link accounts} and {@link accountsPartial} methods for automatically
+   * resolving accounts.
+   */
+  public accountsStrict(accounts: Accounts<A>) {
+    this._resolveAccounts = false;
     this._accountsResolver.resolveOptionals(accounts);
     return this;
   }
 
-  public signers(signers: Array<Signer>): MethodsBuilder<IDL, I> {
+  public signers(signers: Array<Signer>) {
     this._signers = this._signers.concat(signers);
     return this;
   }
 
-  public remainingAccounts(
-    accounts: Array<AccountMeta>
-  ): MethodsBuilder<IDL, I> {
+  public remainingAccounts(accounts: Array<AccountMeta>) {
     this._remainingAccounts = this._remainingAccounts.concat(accounts);
     return this;
   }
 
-  public preInstructions(
-    ixs: Array<TransactionInstruction>
-  ): MethodsBuilder<IDL, I> {
+  public preInstructions(ixs: Array<TransactionInstruction>) {
     this._preInstructions = this._preInstructions.concat(ixs);
     return this;
   }
 
-  public postInstructions(
-    ixs: Array<TransactionInstruction>
-  ): MethodsBuilder<IDL, I> {
+  public postInstructions(ixs: Array<TransactionInstruction>) {
     this._postInstructions = this._postInstructions.concat(ixs);
     return this;
   }
 
+  /**
+   * Get the public keys of the instruction accounts.
+   *
+   * The return type is an object with account names as keys and their public
+   * keys as their values.
+   *
+   * Note that an account key is `undefined` if the account hasn't yet been
+   * specified or resolved.
+   */
+  public async pubkeys(): Promise<
+    Partial<InstructionAccountAddresses<IDL, I>>
+  > {
+    if (this._resolveAccounts) {
+      await this._accountsResolver.resolve();
+    }
+    // @ts-ignore
+    return this._accounts;
+  }
+
   public async rpc(options?: ConfirmOptions): Promise<TransactionSignature> {
-    if (this._autoResolveAccounts) {
+    if (this._resolveAccounts) {
       await this._accountsResolver.resolve();
     }
 
@@ -234,24 +288,23 @@ export class MethodsBuilder<IDL extends Idl, I extends AllInstructions<IDL>> {
       remainingAccounts: this._remainingAccounts,
       preInstructions: this._preInstructions,
       postInstructions: this._postInstructions,
-      options: options,
-      discriminator: this._discriminator,
+      options,
     });
   }
 
   public async rpcAndKeys(options?: ConfirmOptions): Promise<{
-    pubkeys: Partial<InstructionAccountAddresses<IDL, I>>;
+    pubkeys: InstructionAccountAddresses<IDL, I>;
     signature: TransactionSignature;
   }> {
     const pubkeys = await this.pubkeys();
     return {
-      pubkeys,
+      pubkeys: pubkeys as Required<InstructionAccountAddresses<IDL, I>>,
       signature: await this.rpc(options),
     };
   }
 
   public async view(options?: ConfirmOptions): Promise<any> {
-    if (this._autoResolveAccounts) {
+    if (this._resolveAccounts) {
       await this._accountsResolver.resolve();
     }
 
@@ -266,14 +319,14 @@ export class MethodsBuilder<IDL extends Idl, I extends AllInstructions<IDL>> {
       remainingAccounts: this._remainingAccounts,
       preInstructions: this._preInstructions,
       postInstructions: this._postInstructions,
-      options: options,
+      options,
     });
   }
 
   public async simulate(
     options?: ConfirmOptions
   ): Promise<SimulateResponse<any, any>> {
-    if (this._autoResolveAccounts) {
+    if (this._resolveAccounts) {
       await this._accountsResolver.resolve();
     }
 
@@ -284,12 +337,12 @@ export class MethodsBuilder<IDL extends Idl, I extends AllInstructions<IDL>> {
       remainingAccounts: this._remainingAccounts,
       preInstructions: this._preInstructions,
       postInstructions: this._postInstructions,
-      options: options,
+      options,
     });
   }
 
   public async instruction(): Promise<TransactionInstruction> {
-    if (this._autoResolveAccounts) {
+    if (this._resolveAccounts) {
       await this._accountsResolver.resolve();
     }
 
@@ -300,13 +353,15 @@ export class MethodsBuilder<IDL extends Idl, I extends AllInstructions<IDL>> {
       remainingAccounts: this._remainingAccounts,
       preInstructions: this._preInstructions,
       postInstructions: this._postInstructions,
-      discriminator: this._discriminator,
     });
   }
 
   /**
-   * Convenient shortcut to get instructions and pubkeys via
-   * const { pubkeys, instructions } = await prepare();
+   * Convenient shortcut to get instructions and pubkeys via:
+   *
+   * ```ts
+   * const { pubkeys, instructions } = await method.prepare();
+   * ```
    */
   public async prepare(): Promise<{
     pubkeys: Partial<InstructionAccountAddresses<IDL, I>>;
@@ -316,12 +371,12 @@ export class MethodsBuilder<IDL extends Idl, I extends AllInstructions<IDL>> {
     return {
       instruction: await this.instruction(),
       pubkeys: await this.pubkeys(),
-      signers: await this._signers,
+      signers: this._signers,
     };
   }
 
   public async transaction(): Promise<Transaction> {
-    if (this._autoResolveAccounts) {
+    if (this._resolveAccounts) {
       await this._accountsResolver.resolve();
     }
 
@@ -332,7 +387,6 @@ export class MethodsBuilder<IDL extends Idl, I extends AllInstructions<IDL>> {
       remainingAccounts: this._remainingAccounts,
       preInstructions: this._preInstructions,
       postInstructions: this._postInstructions,
-      discriminator: this._discriminator,
     });
   }
 }